insta-v1/requirements.txt
asgiref==3.4.1
crispy==0.6.3
cycler==0.10.0
Django==3.2.7
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-redis==5.0.0
djangorestframework==3.12.4
fabio==0.12.0
gunicorn==20.1.0
h5py==3.4.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.2
psycopg2-binary==2.9.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
silx==0.15.2
six==1.16.0
sqlparse==0.4.2
